+++ tests/test_events_with_assembly.py
@@ -12,7 +12,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')
 
 from csw.Coords import EqCoord, EqFrame, SolarSystemCoord, SolarSystemObject, MinorPlanetCoord, \
     CometCoord, AltAzCoord
-from csw.Parameter import Parameter, Struct
+from csw.Parameter import Parameter
 from csw.Event import SystemEvent
 from csw.EventPublisher import EventPublisher
 
@@ -117,12 +117,8 @@ class TestEventsWithAssembly:
         coordsParam = Parameter("CoordParam", "CoordKey",
                                 [eqCoord, solarSystemCoord, minorPlanetCoord, cometCoord, altAzCoord])
 
-        structParam = Parameter("MyStruct", "StructKey", [Struct(
-            [coordsParam, intParam, floatParam, longParam, shortParam, booleanParam, intArrayParam, floatArrayParam,
-             doubleArrayParam, intMatrixParam])])
-
         paramSet = [coordsParam, byteParam, intParam, floatParam, longParam, shortParam, booleanParam, byteArrayParam,
-                    intArrayParam, floatArrayParam, doubleArrayParam, intMatrixParam, structParam]
+                    intArrayParam, floatArrayParam, doubleArrayParam, intMatrixParam]
         event = SystemEvent(self.prefix, "testEvent3", paramSet)
         self.pub.publish(event)